Seperated the method of validation into new class
